Add history controller

+++ sport-shop-api/Controllers/ProductsController.cs
@@ -26,6 +26,7 @@ namespace sport_shop_api.Controllers
         public async Task<ActionResult<IEnumerable<Product>>> GetProducts()
         {
             List<Product> products = await _context.Products.Include(p => p.Sizes).ToListAsync();
+
             return Ok(products);
         }
 
@@ -33,44 +34,45 @@ namespace sport_shop_api.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> PutProduct(int id, [FromForm] ProductFileDTO newProductDTO)
         {
-            newProductDTO.Id = id;
-
             try
             {
-                Product oldProduct = await _context.Products.FindAsync(newProductDTO.Id);
-                Product newProduct = _mapper.Map<Product>(newProductDTO);
+                Product oldProduct = await _context.Products.FindAsync(id);
 
                 if (newProductDTO.File?.Length > 0)
                 {
                     string stroredPath = Path.GetFullPath(Path.Combine(Environment.CurrentDirectory, "wwwroot/files"));
-                    string imgPath = oldProduct.Url[47..];
-                    string fullPath = Path.Combine(stroredPath, imgPath);
-                    System.IO.File.Delete(fullPath);
+
+                    if (oldProduct.Url.Contains("azurewebsites"))
+                    {
+                        string imgPath = oldProduct.Url[47..];
+                        string fullPath = Path.Combine(stroredPath, imgPath);
+                        System.IO.File.Delete(fullPath);
+                    }
 
                     string newImgPath = DateTime.Now.ToString("yyyyMMddTHHmmss") + newProductDTO.File.FileName;
-                    string newFullPath = Path.Combine(stroredPath, imgPath);
-                    using (var fileStream = new FileStream(fullPath, FileMode.Create))
+                    string newFullPath = Path.Combine(stroredPath, newImgPath);
+                    using (var fileStream = new FileStream(newFullPath, FileMode.Create))
                     {
                         await newProductDTO.File.CopyToAsync(fileStream);
                     }
                     string Url = $@"https://sport-shop-api.azurewebsites.net/files/{newImgPath}";
-                    newProduct.Url = Url;
-                    _context.Entry(newProduct).State = EntityState.Detached;
-                    await _context.SaveChangesAsync();
-                    return Ok(new { Url });
-                }
-                else
-                {
-                    newProduct.Url = oldProduct.Url;
-                    _context.Entry(newProduct).State = EntityState.Detached;
-                    await _context.SaveChangesAsync();
-                    return Ok();
-                }
+                    oldProduct.Url = Url;
 
+                }
+                oldProduct.Name = newProductDTO.Name;
+                oldProduct.Description = newProductDTO.Description;
+                oldProduct.Quantity = newProductDTO.Quantity;
+                oldProduct.Price = newProductDTO.Price;
+                oldProduct.CategoryId = newProductDTO.CategoryId;
+
+                _context.Entry(oldProduct).State = EntityState.Modified;
+
+                await _context.SaveChangesAsync();
+                return Ok(new { oldProduct.Url });
             }
             catch (Exception)
             {
-                return BadRequest();
+                return BadRequest("dasd");
             }
         }
 
@@ -80,6 +82,7 @@ namespace sport_shop_api.Controllers
         {
             try
             {
+                Product product = _mapper.Map<Product>(productDTO);
                 if (productDTO.File?.Length > 0)
                 {
                     string stroredPath = Path.GetFullPath(Path.Combine(Environment.CurrentDirectory, "wwwroot/files"));
@@ -95,15 +98,18 @@ namespace sport_shop_api.Controllers
                     }
                     string Url = $@"https://sport-shop-api.azurewebsites.net/files/{imgPath}";
 
-                    Product product = _mapper.Map<Product>(productDTO);
                     product.Url = Url;
-                    _context.Products.Add(product);
-                    await _context.SaveChangesAsync();
-
-                    return Ok(new { Url });
                 }
+                product.Sizes = new List<ProductSize>
+                {
+                    new() { Name = "S", Price = product.Price },
+                    new() { Name = "M", Price = product.Price },
+                    new() { Name = "L", Price = product.Price },
+                };
+                _context.Products.Add(product);
+                await _context.SaveChangesAsync();
 
-                return BadRequest("File not found");
+                return Ok(new { product.Url, product.Id });
             }
 
             catch (Exception)
@@ -122,12 +128,13 @@ namespace sport_shop_api.Controllers
             {
                 return NotFound();
             }
-
-            string stroredPath = Path.GetFullPath(Path.Combine(Environment.CurrentDirectory, "wwwroot/files"));
-            string imgPath = product.Url[47..];
-            string fullPath = Path.Combine(stroredPath, imgPath);
-            System.IO.File.Delete(fullPath);
-
+            if (product.Url.Contains("azurewebsites"))
+            {
+                string stroredPath = Path.GetFullPath(Path.Combine(Environment.CurrentDirectory, "wwwroot/files"));
+                string imgPath = product.Url[47..];
+                string fullPath = Path.Combine(stroredPath, imgPath);
+                System.IO.File.Delete(fullPath);
+            }
             _context.Products.Remove(product);
             await _context.SaveChangesAsync();
             return Ok();
